refactor: split module handlers
